[Chore] Storybook 샘플 컴포넌트 경로 변경 및 __stories__ 구조 적용 (#3)
- src/stories 샘플 컴포넌트 제거
- __stories__ 폴더를 components/ui 내부에 적용
- main.ts stories 경로 수정
